refactor frontend setting config
