--- web/src/constants/settings/identity.ts
@@ -1,10 +1,0 @@
-export const identitySetting = {
-  usernameMinLength: 3,
-  usernameMaxLength: 20,
-  usernameAllowedChars: "abcdefghijklmnopqrstuvwxyz-_",
-  passwordMinimumLength: 6,
-  passwordRequiresDigit: true,
-  passwordRequiresLowercase: true,
-  passwordRequiresUppercase: true,
-  passwordRequiresNonAlphanumeric: false
-}
