--- web/src/constants/settings/index.ts
@@ -1,10 +1,0 @@
-import { identitySetting } from "./identity";
-import { audioUploadSetting, imageUploadSetting } from "./upload";
-
-export default {
-  IDENTITY: identitySetting,
-  UPLOAD: {
-    AUDIO: audioUploadSetting,
-    IMAGE: imageUploadSetting
-  }
-}
