--- web/src/constants/settings/upload.ts
@@ -1,14 +1,0 @@
-type UploadSettingType = {
-  accept: string[],
-  maxSize: number
-}
-
-export const audioUploadSetting: UploadSettingType = {
-  accept: ["audio/mp3", "audio/mpeg"],
-  maxSize: 262144000
-}
-
-export const imageUploadSetting: UploadSettingType = {
-  accept: ["image/jpeg", "image/png", "image/gif"],
-  maxSize: 2097152
-}
